src/requirements.txt
requests==2.31.0
tiktoken==0.6.0
tabulate==0.9.0
python-dotenv==1.0.1